requirements.txt
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
scikit-learn~=1.6.0
python-docx~=1.1.2